requirements.txt
tensorflow
numpy
matplotlib
tensorflow-examples @ git+https://github.com/tensorflow/examples.git
Pillow
pydot
graphviz
keras
ipython
jinja2
tensorflow-model-optimization
flask
gunicorn 